requirements.txt
fastapi>=0.108.0
uvicorn>=0.27.0
aiohttp>=3.9.1
python-dotenv>=1.0.0
loguru>=0.7.2
requests>=2.31.0
Pillow>=10.1.0
cloudinary>=1.36.0
beautifulsoup4>=4.12.2
python-multipart>=0.0.6
websockets>=12.0
typing-extensions>=4.8.0
httpx>=0.25.2
